Update - Change the name of s class and added another one
I changed the div of class card-img to class card-img-1 since it is under the first section.
I added a div of class card-img-2 containing the image in the second section.
For the main.css, I also changed the styles applied to the section of id sect-1 so that it will be applied to all sections and I changed its padding bottom.